[server] Factor out mutex map
Factors out a data structure that tracks lockable resources by key.
